+++ Bugsnag/Payload/BugsnagStackframe+Private.h
@@ -14,7 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSArray<BugsnagStackframe *> *)stackframesWithBacktrace:(uintptr_t *)backtrace length:(NSUInteger)length;
 
-/// Constructs a stackframe object from a stackframe dictionary and list of images captured by KSCrash.
+/// Constructs a stackframe object from a KSCrashReport backtrace dictionary.
 + (nullable instancetype)frameFromDict:(NSDictionary<NSString *, id> *)dict withImages:(NSArray<NSDictionary<NSString *, id> *> *)binaryImages;
 
 /// Constructs a stackframe object from a JSON object (typically loaded from disk.)